requirements.txt
python-dotenv
pydantic~=2.11.5
jsonschema~=4.23.0
email-validator

requests>=2.32.0
pytest>=7.4.0
fastapi~=0.115.12
uvicorn~=0.34.2

fastapi-pagination
psycopg2-binary
sqlalchemy
sqlmodel

faker